+++ tests/test_attributes.py
@@ -3,79 +3,81 @@ import pytest
 from confu.schema import (
     Schema,
     Attribute,
-    StringAttribute,
-    IntAttribute,
-    FloatAttribute,
-    BoolAttribute,
-    ListAttribute,
-    DirectoryAttribute
+    Str,
+    Int,
+    Float,
+    Bool,
+    List,
+    Directory
 )
 
 
 
-@pytest.mark.parametrize("AttributeClass", [
-    StringAttribute,
-    IntAttribute,
-    FloatAttribute,
-    BoolAttribute,
-    DirectoryAttribute])
-def test_init(AttributeClass):
+@pytest.mark.parametrize("Class", [
+    Str,
+    Int,
+    Float,
+    Bool,
+    Directory])
+def test_init(Class):
     """
     test attribute initiation
     """
-    attribute = AttributeClass("test", help="test attribute")
+    attribute = Class("test", help="test attribute")
     assert attribute.name == "test"
     assert attribute.help == "test attribute"
 
 
-@pytest.mark.parametrize("AttributeClass,choices_in,choices_out", [
-    (StringAttribute, ["a","b"], ["a","b"]),
-    (StringAttribute, lambda x: ["a","b"], ["a","b"]),
-    (IntAttribute, [1,2], [1,2]),
-    (IntAttribute, lambda x: [1,2], [1,2]),
-    (FloatAttribute, [1.2,2.3], [1.2,2.3]),
-    (FloatAttribute, lambda x: [1.2,2.3], [1.2,2.3]),
-    (DirectoryAttribute, ["/a", "/b"], ["/a", "/b"]),
-    (DirectoryAttribute, lambda x: ["/a","/b"], ["/a","/b"])
+@pytest.mark.parametrize("Class,choices_in,choices_out", [
+    (Str, ["a","b"], ["a","b"]),
+    (Str, lambda x: ["a","b"], ["a","b"]),
+    (Int, [1,2], [1,2]),
+    (Int, lambda x: [1,2], [1,2]),
+    (Float, [1.2,2.3], [1.2,2.3]),
+    (Float, lambda x: [1.2,2.3], [1.2,2.3]),
+    (Directory, ["/a", "/b"], ["/a", "/b"]),
+    (Directory, lambda x: ["/a","/b"], ["/a","/b"])
 ])
-def test_choices(AttributeClass, choices_in, choices_out):
+def test_choices(Class, choices_in, choices_out):
     """
     test attribute choices property
     """
-    attribute = AttributeClass("test", choices=choices_in)
+    attribute = Class("test", choices=choices_in)
     assert attribute.choices == choices_out
 
 
-@pytest.mark.parametrize("AttributeClass,default_in,default_out", [
-    (StringAttribute, "test", "test"),
-    (StringAttribute, lambda x: "test", "test"),
-    (IntAttribute, 123, 123),
-    (IntAttribute, lambda x: 123, 123),
-    (FloatAttribute, 1.23, 1.23),
-    (FloatAttribute, lambda x: 1.23, 1.23),
-    (BoolAttribute, True, True),
-    (BoolAttribute, lambda x: True, True),
-    (DirectoryAttribute, "/test", "/test"),
-    (DirectoryAttribute, lambda x: "/test", "/test"),
+@pytest.mark.parametrize("Class,default_in,default_out", [
+    (Str, "test", "test"),
+    (Str, lambda x: "test", "test"),
+    (Int, 123, 123),
+    (Int, lambda x: 123, 123),
+    (Float, 1.23, 1.23),
+    (Float, lambda x: 1.23, 1.23),
+    (Bool, True, True),
+    (Bool, lambda x: True, True),
+    (Directory, "/test", "/test"),
+    (Directory, lambda x: "/test", "/test"),
 ])
-def test_default(AttributeClass, default_in, default_out):
+def test_default(Class, default_in, default_out):
     """
     test attribute default property
     """
-    attribute = AttributeClass("test", default=default_in)
+    attribute = Class("test", default=default_in)
     assert attribute.default == default_out
 
 
 def test_list():
     with pytest.raises(TypeError):
-        attribute = ListAttribute("test", 123)
+        attribute = List("test", 123)
 
-    item = IntAttribute("item")
-    attribute = ListAttribute("test", item)
+    item = Int("item")
+    attribute = List("test", item)
     assert attribute.item == item
     assert attribute.name == "test"
 
     class Item(Schema):
-        test_sub = IntAttribute("test_sub")
-    attribute = ListAttribute("test", Item)
+        test_sub = Int("test_sub")
+    attribute = List("test", Item)
     assert attribute.item == Item
+
+    attribute = List("test", Int())